wrap: Extract yaml into separated module.
As in commit subject + provide Format interface.
